Move stat-related methods of Event to EventStatTrait for readability
